1) System V AMD64 ABI calling convention is: the first six integer or pointer arguments are passed in registers the remaining arguments are pushed on stack.
2) "char" and "short" are promoted to "int" (32-bit on AMD64) when passed as varargs

3) %P and %d expected 64-bit integer before

So 8, 16 and 32-bit integer arguments (starting from 4th or 5th)
for rprintf() and friends are not printed properly.

Fix this by the following:
- %d now expects "int"
- implement l-modifier (%ld) to print 64-bit int
- implement %x (hex-formatted int) and %lx
- implement width modifier for %d and %x (%02x)
only zero-padding (no space-padding) is implemented for now
- move %s to common runtime
- remove %P (%lx should be used instead)
- remove %E (use %s and strerror(errno) instead)
- format_number: fix printing 0 when pad == 0 (some callers used pad == 0)
- vbprintf() tests

Note that %ld in 32-bit code (-DBITS32) also expects 64-bit (not "long").
We never use "long" and always use "u64" or "s64" explicitly.

+++ src/gdb/gdbstub.c
@@ -24,8 +24,8 @@ static void reset_parser(gdb g)
 
 static context gdb_handle_exception (gdb g, context frame)
 {
-    int exceptionVector = frame[FRAME_VECTOR];
-    //     rprintf ("gdb exception: %d %p [%p %p] %p %p\n", exceptionVector, g, frame, g->t->frame, frame[FRAME_RIP], *(u64 *)frame[FRAME_RIP]);
+    u64 exceptionVector = frame[FRAME_VECTOR];
+    //     rprintf ("gdb exception: %ld %p [%p %p] %p %p\n", exceptionVector, g, frame, g->t->frame, frame[FRAME_RIP], *(u64 *)frame[FRAME_RIP]);
     sigval = computeSignal(exceptionVector);
     reset_buffer(g->output);
     /*
